Refactoring: utils/useMobile
Introduce useMobile util to handle the mobile checks in one place.

Also introduce an export-index file so we can import all utils in one line.
