+++ src/utils/useMobile.js
@@ -1,0 +1,4 @@
+// Very, very basic mobile detection.
+export const useMobile = () => {
+  return window.innerWidth <= 800;
+};
