+++ src/views/teams.js
@@ -6,7 +6,7 @@ import { push } from 'react-router-redux';
 
 import { modal } from '../store/modal_actions';
 import { logUserOut } from '../store/auth_actions';
-import { cancelablePromise } from '../utils/promise';
+import { cancelablePromise, useMobile } from '../utils';
 import {
   createMappingTeam,
   fetchUserMappingTeams,
@@ -135,10 +135,12 @@ class MappingTeams extends React.PureComponent<any, propsType, any> {
   };
 
   render() {
+    const mobile = useMobile();
+
     return (
       <div
         className={`flex-parent flex-parent--column changesets-filters bg-white${
-          window.innerWidth < 800 ? 'viewport-full' : ''
+          mobile ? 'viewport-full' : ''
         }`}
       >
         <header className="h55 hmin55 flex-parent px30 bg-gray-faint flex-parent--center-cross justify--space-between color-gray border-b border--gray-light border--1">
